+++ hue-api/utils.js
@@ -6,7 +6,7 @@ const HueError = require('../lib/HueError');
 //TODO remove this module or refactor
 
 module.exports.nativePromiseOrCallback = function (promise, cb) {
-  var promiseResult = promise;
+  let promiseResult = promise;
 
   if (cb && typeof cb === "function") {
     module.exports.resolvePromise(promise, cb);